Moved routablepage test model into its own app
