+++ wagtail/tests/routablepage/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class WagtailRoutablePageTestsAppConfig(AppConfig):
+    name = 'wagtail.tests.routablepage'
+    label = 'routablepagetests'
+    verbose_name = "Wagtail routable page tests"
